Writing a template generator in Lua (llgen.lua).
First try at writing a a Lua script to parse Leptonica's allheaders.h
and generate lualept-style C function wrappers from it.

The output is a file `template.cpp' and can be used to copy+paste
code to the src/ll*.cpp files. Manual inspection and correction is
required in any case.