A lot of cleaning up code and refacturing ll_new_*.
To keep the Lua class "Create" functions clean, do parameter interpretation
in the ll_create_<ClassName> functions instead.

This is meant to make it possible to e.g. call the constructor:
	Box(0,0,32,32)
or, if the Arg#1 is another Box*
	Box(boxs)
to automagically call Leptonica's boxCopy(boxs) to construct the
new Box*.

Also change the meaning of debug enables to a bitmask, which will
later be settable during runtine to enable / disable debugging of
specific sections of the code.

Finally moved the image(s) to ./images and adjusted the script
and Doxyfile.