Fix debug code; rename DisplaySDL2() and improve.
The debug code was broken, now it is fixed and compiles.

Rename DisplaySDL2() to ShowSDL2() and add a Pix.Show(...) C function.

Not all formats can be displayed yet and rendering the Pix* onto
the SDL2_Surface is not optimal. Ideally we would use
SDL_CreateSurfaceFrom() with parameters matching the memory layout
of the Pix* to display.