Add a X11 color lookup table; misc. fixes.
Add a new function ll_unpack_Matrix() which accepts a 2D w, h matrix
as either a 2D table array, a flattened 1D table array, or as a sequence
of arguments (numbers) for the total of w * h values.
Test this with Box:AffineTransform().

Add a table of color names and their RGB values and the corresponding
functions ll_check_color_name() and ll_string_color_name().
The latter returns either a name or, if none is defined for the color,
a string of the format "#RRGGBB" where RR, GG and BB are the hexadecimal
values of the color's red, green and blue components.