Add code to set/get Lua globals.
In a first step to support passing variables between the application
and Lua, lualept now defines an interface to add a global variable
to set (before running a script) and get (after running a script)
to an internal list.

This list uses the new typedef struct global_var_s which contains
the lualept type name, the variable name, and a pointer to the
variable in the applications memory.

To avoid excessive type casting, the pointers are defined as a
union pointers of all known types, or as pointers to pointers in
case of get global.

This is yet untested in the scripting but that's the next step.