Move program (llua) to prog/; fix handling of globals.
Move the program llua to its own subdirectory. There may be more
examples in the future and they should be sparated from the library.

Add code to simplify the creation of lists of global variables
for the application. Fix ll_set_global() to also expect a pointer
to the variable (i.e. &var) and create an object for this variable's
contents (e.g. Box*).