Fix passing of globals between C and Lua.
The ll_global_<ClassName> (now ll_get_global_<ClassName>) functions
used the wrong argument index 1 instead of -1.

Now it is possible to take "objects" from a Lua user data and giving
it back to C. The pointer to the variable does not have to be
differen for in- and output, so you can simply use the same array
of ll_global_var_t for set and get globals.

The prog/llua.cpp sample code demonstrates how to use this
interface.

I could instead have used the return values from main, but you
would write one more Lua statement and would have had to supply the
pointers to where the Lua variables are to be saved away in any case.