Refactor get/set globals; fix color value args.
It was a bad idea to have a global variable for the array of
set and get global Lua variables. Make the functions private
to lualept.cpp and refactor them to use a caller specified
root "global_vars_t *&gvars" so that they live for only
as long as the script is run.

Rename ll_RunScript() to just ll_run() and make it execute
either a script from a file (filename is given, script is nullptr)
or a string (filename is the name of the script, script is a string).

Also pass the application's set_vars and get_vars arrays to
ll_run() so that it can construct the gvars locally.

Create a new function ll_check_color_index() which tries to deduce
whether the argument is a color value or pixel index by inspecting
the depth of the optionally supplied Pix* pix.