Refactor Destroy C functions using ll_take_udata().
Simplify the Destroy() functions by using a templated
ll_take_udata() to take the Leptonica class pointer
from the Lua user data and replace it with a nullptr.

Add some missing typedefs for Rbtree and make the llgen.lua
script generate better template code for the Destroy()
functions. TODO: make sure nothing is pushed to the Lua
stack in Destroy() and return 0 from the function.