First try at creating cmake project files.
Taken mostly from Leptonica and adapted to check for the
libraries required for lualept, i.e. leptonica, lua5.3 and
optionally SDL2 for the internal viewer.

Made the ll_unpack_* functions allocate the memory for
the data instead of expecting the caller to do that.

Add debug info for pushing nil (error case).